refactor(RacingGameResultDrawer): can draw car status(name + distance):
- 변경된 요구사항을 충족시키기 위해 차량명과 차량의 이동거리가 모두 사용자에게 보이도록 기능 개선
