+++ README.md
@@ -152,5 +152,5 @@ public class SetTest {
     - [ ] 우승 차량들을 계산할 수 있다.
 - RacingGamRunner
     - [X] 게임 시작 방식을 차량 대수 입력에서 참여할 차량 이름으로 변경한 뒤 기존 기능 그대로 동작하는지 확인
-    - [ ] 레이싱 게임 결과 출력 시 차량 이름도 출력되도록 기능 변경
+    - [X] 레이싱 게임 결과 출력 시 차량 이름도 출력되도록 기능 변경
     - [ ] 우승 차량 결과를 사용자가 볼 수 있도록 기능 추가
